Reorganizing folder structure.
