+++ src/app/app.component.ts
@@ -1,13 +1,13 @@
-import { Component } from '@angular/core';
+import { Component }            from '@angular/core';
 
 import { Angular2TokenService } from "angular2-token";
 
-import { environment } from "../environments/environment";
+import { environment }          from "../environments/environment";
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  selector:     'app-root',
+  templateUrl:  './app.component.html',
+  styleUrls:    ['./app.component.css']
 })
 export class AppComponent {
   constructor(private authToken: Angular2TokenService){
