+++ src/app/app.module.ts
@@ -1,56 +1,52 @@
-import { BrowserModule } from '@angular/platform-browser';
+import { NgModule }                         from '@angular/core';
+import { HttpModule }                       from '@angular/http';
+import { BrowserModule }                    from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http'
 
-import { MaterializeModule } from 'angular2-materialize';
-import { Angular2TokenService } from 'angular2-token';
+import { MaterializeModule }                from 'angular2-materialize';
+import { Angular2TokenService }             from 'angular2-token';
 
-import { routes } from './app.routes';
-import { AppComponent } from './app.component';
-import { StudentsComponent } from './students/students.component';
-import { StudentFormComponent } from './students/student-form/student-form.component';
-import { RegisterService } from './service/register/register.service';
-import { StudentService } from './students/shared/student.service';
-import { AuthService } from "./services/auth.service";
-
-import { LoginComponent } from './login/login.component';
-import { ToolbarComponent } from './toolbar/toolbar.component';
-import { AuthDialogComponent } from './auth-dialog/auth-dialog.component';
-import { LoginFormComponent } from './login-form/login-form.component';
-import { RegisterFormComponent } from './register-form/register-form.component';
-import { ProfileComponent } from './profile/profile.component';
-import { AuthGuard } from "./guards/auth.guard";
-
+import { AppComponent }                     from './app.component';
+import { routes }                           from './app.routes';
+import { AuthGuard }                        from "./guards/auth.guard";
+import { AuthService }                      from "./services/auth.service";
+import { AuthDialogComponent }              from './login/auth-dialog/auth-dialog.component';
+import { LoginComponent }                   from './login/login.component';
+import { LoginFormComponent }               from './login/login-form/login-form.component';
+import { RegisterService }                  from './login/shared/register.service';
+import { RegisterFormComponent }            from './login/register-form/register-form.component';
+import { ProfileComponent }                 from './profile/profile.component';
+import { StudentService }                   from './students/shared/student.service';
+import { StudentsComponent }                from './students/students.component';
+import { StudentFormComponent }             from './students/student-form/student-form.component';
+import { ToolbarComponent }                 from './toolbar/toolbar.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    StudentsComponent,
-    StudentFormComponent,
-    LoginComponent,
-    ToolbarComponent,
     AuthDialogComponent,
+    LoginComponent,
     LoginFormComponent,
     RegisterFormComponent,
-    ProfileComponent
+    ProfileComponent,
+    ToolbarComponent,
+    StudentsComponent,
+    StudentFormComponent
   ],
   imports: [
-    FormsModule,
     HttpModule,
-    ReactiveFormsModule,
     BrowserModule,
     FormsModule,
-    HttpModule,
+    ReactiveFormsModule,
     MaterializeModule,
     routes
   ],
   providers: [
     AuthGuard,
     AuthService,
-    StudentService,
     RegisterService,
-    Angular2TokenService
+    Angular2TokenService,
+    StudentService
   ],
   bootstrap: [ AppComponent ]
 })
