+++ src/app/app.routes.ts
@@ -1,12 +1,12 @@
-import { ModuleWithProviders } from '@angular/core';
+import { ModuleWithProviders }    from '@angular/core';
 import { Routes, RouterModule }   from '@angular/router';
 
-import { AuthGuard } from "./guards/auth.guard";
+import { AuthGuard }              from "./guards/auth.guard";
 
-import { LoginComponent } from "./login/login.component";
-import { ProfileComponent } from "./profile/profile.component";
-import { StudentsComponent } from './students/students.component';
-import { StudentFormComponent } from './students/student-form/student-form.component';
+import { LoginComponent }         from "./login/login.component";
+import { ProfileComponent }       from "./profile/profile.component";
+import { StudentsComponent }      from './students/students.component';
+import { StudentFormComponent }   from './students/student-form/student-form.component';
 
 const APP_ROUTES: Routes = [
   { path: '',                   component: LoginComponent,        pathMatch: 'full'                           },
