+++ src/app/toolbar/toolbar.component.ts
@@ -1,17 +1,21 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { AuthDialogComponent } from "../auth-dialog/auth-dialog.component";
-import { AuthService } from "../services/auth.service";
 import { Router } from "@angular/router";
 
+import { AuthService } from "../services/auth.service";
+import { AuthDialogComponent } from "../login/auth-dialog/auth-dialog.component";
+
 @Component({
-  selector: 'app-toolbar',
-  templateUrl: './toolbar.component.html',
-  styleUrls: ['./toolbar.component.css']
+  selector:     'app-toolbar',
+  templateUrl:  './toolbar.component.html',
+  styleUrls:    ['./toolbar.component.css']
 })
 export class ToolbarComponent implements OnInit {
   @ViewChild('authDialog') authDialog: AuthDialogComponent;
 
-  constructor(public authService:AuthService, private router:Router) { }
+  constructor(
+    public authService: AuthService,
+    private router: Router
+  ) { }
 
   ngOnInit() {}
